feat(store): 添加PgVector存储支持并优化检索功能
- 集成langchain4j-pgvector依赖库
- 实现PgVectorEmbeddingStoreImpl存储组件
- 优化检索结果分数类型从Float到Double的转换
- 重构搜索逻辑以提高匹配精度
- 删除已废弃的chatModelId字段
- 添加noActiveList方法用于非活跃段落查询
- 调整向量存储实现中的排序和去重算法
- 更新系统权限接口实现以优化性能

+++ maxkb4j-service-api/maxkb4j-knowledge-api/src/main/java/com/maxkb4j/knowledge/retrieval/SearchResult.java
@@ -31,12 +31,12 @@ public class SearchResult {
     /**
      * Maximum similarity score in results
      */
-    private float maxScore;
+    private double maxScore;
 
     /**
      * Minimum similarity score in results
      */
-    private float minScore;
+    private double minScore;
 
     /**
      * Time taken for search in milliseconds
@@ -64,11 +64,11 @@ public class SearchResult {
             return empty();
         }
 
-        float max = Float.MIN_VALUE;
-        float min = Float.MAX_VALUE;
+        double max = Double.MIN_VALUE;
+        double min = Double.MAX_VALUE;
 
         for (TextChunkVO chunk : chunks) {
-            float score = chunk.getScore() != null ? chunk.getScore() : 0.0f;
+            double score = chunk.getScore() != null ? chunk.getScore() : 0d;
             if (score > max) max = score;
             if (score < min) min = score;
         }
